refactor: rename transfer-options to transfer and update imports
Simplify file naming and fix import paths for consistency.

#VERCEL_SKIP
