+++ app/wallet/components/index.ts
@@ -1,2 +1,2 @@
 export { default as TransactionsTab } from "./transactions-tab"
-export { default as TransferOptions } from "./transfer-options"
+export { default as TransferOptions } from "./transfer"
